requirements.txt
flask
flask-cors
tensorflow==2.15.0
numpy
pillow
gunicorn
